+++ sample_projects/template3D/custom_modules/custom.cpp
@@ -230,7 +230,7 @@ void setup_tissue( void )
 
 std::vector<std::string> my_coloring_function( Cell* pCell )
 {
-	// start with the Ki67 coloring 
+	// start with flow cytometry coloring 
 	
 	std::vector<std::string> output = false_cell_coloring_cytometry(pCell);