+++ sample_projects/template3D/main-3D.cpp
@@ -208,6 +208,10 @@ int main( int argc, char* argv[] )
 			// run PhysiCell 
 			((Cell_Container *)microenvironment.agent_container)->update_all_cells( PhysiCell_globals.current_time );
 			
+			/*
+			  Custom add-ons could potentially go here. 
+			*/			
+			
 			PhysiCell_globals.current_time += diffusion_dt;
 		}