+++ changes.md
@@ -75,6 +75,8 @@ Minor new features and changes:
   
   4) Updated to use the my_coloring_function coloring function. Made sure the my_coloring_function used false_cell_coloring_cytometry as its starting point. 
   
++ Updated biorobots sample project:   
+  
   
 + New function in ./core/PhysiCell_utilities/*: